Improvement part 1 + WIP part 2
An improvement was done in the part 1 (type of the response) for the data formatting but maybe there is something better.

In the part 2, the call API is done when we are clicking on the checkbox of a specific ingredient. This call API should be done when we are clicking on the validate button.
The display of cocktails is really not good (horizontal scrollbar). We should fix this.

After fixing these little problems, let's dive into the bonus point and the display of cocktail cards with all their data.