Part 3 - 3.2 - Favorite list ingredients
The first approach was to handle the list of ingredients directly when we add or remove a cocktail from the favorite list. However, the ingredients are already part of the cocktails state. We then create a custom memoize selector to retrieve ingredients data directly from the cocktails state. Let's handle the bonus point with the alcohol-related filter.