Part 4 - Bonus point - alcohol filter - WIP
Many solutions could have been used to manage this alcohol filter.

First of all, we could have fetch throught the API all cocktails that are with [URL?a=Alcoholic] or without [URL?a=Non_Alcoholic] Alcohol. After this, we could have picked in this list all our filtered cocktails by ingredients if they exist.
This solution is feasible but seems not optimized because we make an API call to retrieve the alcohol data even though this data is already present in the retrieved cocktails.

Thus, the second solution is to filter the existing cocktails that we retrieved depending on the value of the strAlcoholic property (=== || !== "Alcoholic"). We used the redux state to have a screenshot of the filtered cocktails because we don't want to override permanently the state of the FindCocktail component if we are changing the value of the alcohol filter. We chose this solution.

The last thing to do to have a functional prototype is to replace <Checkbox /> component by <Radio /> component. In fact, we don't want to manage multiple ingredients into the filter ingredients list. Instead of using <Checkbox /> and fixing bugs when we are select/deselect multiple filter ingredients, we could simply use the <Radio /> component because they are mutually exclusive.